Refactor `wordchipper-cli` with modular structure for readability and maintainability
- Split `common.rs` into smaller files: `disk_cache.rs`, `model_selector.rs`, `tokenizer_mode.rs`, and `input_output.rs`.
- Introduced submodule `commands` for `cat` and `list_models` commands.
- Updated main CLI logic to use new modular layout.
- Simplified argument handling and moved logic into respective components.
