--- crates/wordchipper-cli/src/list_models_command.rs
@@ -1,18 +1,0 @@
-/// Args for the model listing command.
-#[derive(clap::Args, Debug)]
-pub struct ListModelsArgs {}
-
-pub fn run_list_models(_args: &ListModelsArgs) -> Result<(), Box<dyn std::error::Error>> {
-    let vocabs = wordchipper::list_vocabs();
-
-    for listing in &vocabs {
-        println!("\"{}\" - {}", listing.source, listing.description);
-
-        for vocab in &listing.vocabs {
-            println!("  * \"{}:{}\"", listing.source, vocab.id);
-            println!("    {}", vocab.description);
-        }
-    }
-
-    Ok(())
-}
