+++ crates/wordchipper-cli/src/main.rs
@@ -1,11 +1,11 @@
-pub mod cat_command;
-pub mod common;
-pub mod list_models_command;
+mod commands;
+mod disk_cache;
+mod input_output;
+mod model_selector;
+mod tokenizer_mode;
 
-use cat_command::CatArgs;
 use clap::Parser;
-
-use crate::list_models_command::ListModelsArgs;
+use commands::Commands;
 
 /// wordchipper-cli
 #[derive(clap::Parser, Debug)]
@@ -13,30 +13,9 @@ pub struct Args {
     /// Subcommand to run.
     #[clap(subcommand)]
     pub command: Commands,
-
-    #[clap(flatten)]
-    pub disk_cache: common::DiskCacheArgs,
-}
-
-/// Subcommands for wordchipper-cli
-#[derive(clap::Subcommand, Debug)]
-pub enum Commands {
-    /// Act as a streaming tokenizer.
-    Cat(CatArgs),
-
-    /// List available models.
-    ListModels(ListModelsArgs),
 }
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
     let args = Args::parse();
-
-    match &args.command {
-        Commands::Cat(cat_args) => cat_command::run_cat(&args, cat_args)?,
-        Commands::ListModels(list_models_args) => {
-            list_models_command::run_list_models(list_models_args)?
-        }
-    }
-
-    Ok(())
+    args.command.run()
 }
